+++ packages/dso-toolkit/src/components/header/header.args.ts
@@ -17,6 +17,7 @@ export interface HeaderArgs {
   authStatus: "none" | "loggedOut" | "loggedIn";
   loginUrl: string;
   logoutUrl: string;
+  helpUrl: string;
   userProfileName: string;
   userProfileUrl: string;
   userHomeUrl: string;
@@ -56,6 +57,12 @@ export const headerArgTypes: ArgTypes<HeaderArgs> = {
   },
   logoutUrl: {
     name: "Uitloggen url",
+    control: {
+      type: "text",
+    },
+  },
+  helpUrl: {
+    name: "Help url",
     control: {
       type: "text",
     },