+++ storybook/src/example-pages/toepassingen/regels-op-de-kaart/locatie-overzicht.content.ts
@@ -29,6 +29,7 @@ export const header: Header = {
   authStatus: "loggedIn",
   loginUrl: "#login",
   logoutUrl: "#logout",
+  helpUrl: "#help",
   userProfileName: "J.A. Jansen",
   userProfileUrl: "#profiel",
 };